rpk/cli: Add 'rpk topic info', depr. 'rpk api topic status'
